rename DevCompilerBootstrapBuilder to WebEntrypointBuilder and move ddc logic to a separate file
