+++ build_web_compilers/lib/build_web_compilers.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/dev_compiler_bootstrap_builder.dart'
-    show DevCompilerBootstrapBuilder, bootstrapJsExtension;
 export 'src/dev_compiler_builder.dart'
     show
         DevCompilerBuilder,
@@ -17,3 +15,5 @@ export 'src/summary_builder.dart'
         UnlinkedSummaryBuilder,
         linkedSummaryExtension,
         unlinkedSummaryExtension;
+export 'src/web_entrypoint_builder.dart'
+    show WebEntrypointBuilder, ddcBootstrapExtension;
